+++ tests/examples/hashing/AssemblyInfo.fs
@@ -4,20 +4,20 @@ open System.Reflection
 
 [<assembly: AssemblyTitleAttribute("examples")>]
 [<assembly: AssemblyProductAttribute("NpgsqlFSharpAnalyzer")>]
-[<assembly: AssemblyVersionAttribute("1.8.0")>]
-[<assembly: AssemblyMetadataAttribute("ReleaseDate","2020-02-19T00:00:00.0000000")>]
-[<assembly: AssemblyFileVersionAttribute("1.8.0")>]
-[<assembly: AssemblyInformationalVersionAttribute("1.8.0")>]
+[<assembly: AssemblyVersionAttribute("1.9.0")>]
+[<assembly: AssemblyMetadataAttribute("ReleaseDate","2020-02-20T00:00:00.0000000")>]
+[<assembly: AssemblyFileVersionAttribute("1.9.0")>]
+[<assembly: AssemblyInformationalVersionAttribute("1.9.0")>]
 [<assembly: AssemblyMetadataAttribute("ReleaseChannel","release")>]
-[<assembly: AssemblyMetadataAttribute("GitHash","c0489eeb81e7f2240b051c1d2925d89b78136398")>]
+[<assembly: AssemblyMetadataAttribute("GitHash","975c7bd779306059d45b972efaf5c7c17b908fb2")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "examples"
     let [<Literal>] AssemblyProduct = "NpgsqlFSharpAnalyzer"
-    let [<Literal>] AssemblyVersion = "1.8.0"
-    let [<Literal>] AssemblyMetadata_ReleaseDate = "2020-02-19T00:00:00.0000000"
-    let [<Literal>] AssemblyFileVersion = "1.8.0"
-    let [<Literal>] AssemblyInformationalVersion = "1.8.0"
+    let [<Literal>] AssemblyVersion = "1.9.0"
+    let [<Literal>] AssemblyMetadata_ReleaseDate = "2020-02-20T00:00:00.0000000"
+    let [<Literal>] AssemblyFileVersion = "1.9.0"
+    let [<Literal>] AssemblyInformationalVersion = "1.9.0"
     let [<Literal>] AssemblyMetadata_ReleaseChannel = "release"
-    let [<Literal>] AssemblyMetadata_GitHash = "c0489eeb81e7f2240b051c1d2925d89b78136398"
+    let [<Literal>] AssemblyMetadata_GitHash = "975c7bd779306059d45b972efaf5c7c17b908fb2"